Address @michael-smirnov comments
